refactor(ingest): decompose 3_digest.py into focused acyclic seam modules
Split the 2,804-line digest orchestrator into seven leaf modules and remove
the record_enumerator <-> 3_digest import cycle. 3_digest.py is now a
526-line thin orchestrator (-81%).

New modules (each a verbatim extraction, byte-identical to the original):
- _digest_runner.py     multiprocessing watchdog (subprocess isolation + timeouts)
- _bids_path.py         pure BIDS path-parsing / neuro-file classification
- _source_id.py         dataset-id -> source inference
- _manifest_digest.py   manifest (API-only) record synthesis
- _record_extractor.py  per-record BIDS extraction
- _dataset_metadata.py  dataset-level metadata extraction
- _bids_digest.py       BIDS-filesystem record synthesis

De-circularization: ManifestEnumerator/BIDSFilesystemEnumerator now reach
_manifest_digest/_bids_digest via a lazy import instead of importlib-execing
the digit-prefixed 3_digest.py; _load_digest_module() is deleted. The four
seam modules depend only on leaves, never on 3_digest. 3_digest re-exports
the moved symbols for back-compat (megafunction canary + tests).

Also removes dead code (_json_serializer, _positive_float, duplicate
DEFAULT_DATASET_TIMEOUT_SECONDS) and adds characterization tests for the
previously-uncovered multiprocessing watchdog plus decoupling regression
tests pinning the acyclic invariant in CI.

Zero behavior change: every moved function verified byte-identical to HEAD;
byte-level digest snapshot and megafunction canary stay green.
Tests: 973 passed, 0 failures. ruff clean.
